Same fix

+++ app/app.py
@@ -146,7 +146,7 @@ def business_register():
         min_date = date(1920, 1, 1)
         max_date = date.today()
 
-        pp_path = os.path.join('static', 'default.png')
+        pp_path = "https://storage.googleapis.com/madensell-dc0c4.appspot.com//default.png"
         conn = mysql.connector.connect(**config)
         cursor = conn.cursor()
         cursor.execute('SELECT * FROM users WHERE email = %s', (email,))
@@ -656,7 +656,8 @@ def business_edit_profile():
             address = request.form['address']
             profile_image = request.files['profile_pic']
             if profile_image:
-                delete_picture(user[11])
+                if(user[11] != "https://storage.googleapis.com/madensell-dc0c4.appspot.com//default.png"):
+                    delete_picture(user[11])
                 profile_image = add_picture(profile_image, user[0])
             else:
                 profile_image = user[11]
